Выводит текущий тип поездки в форму
